+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 - Fix custom coloured icons
+- Sort folders before files
 
 # v0.0.140 - 2019-08-12
 - Update icons to match the 1.37.0 GUI (great work by [johnletey](https://github.com/johnletey) - many thanks!)